lib: zstd: Upgrade to latest upstream zstd version 1.5.4
- Syncs latest upstream ZSTD from [1].

- This update retains the following commits:
https://github.com/cyberknight777/dragonheart_kernel_oneplus_sm8150/commit/edc41e9a5d8771f86e8951ff3e6599a9db868ba6 {"lib: zstd: Fix attribute declaration"}
https://github.com/cyberknight777/dragonheart_kernel_oneplus_sm8150/commit/31ef7d2d75e080362b144fba1b60b0802cf04816 {"lib: zstd: include a missing header"}
https://github.com/cyberknight777/dragonheart_kernel_oneplus_sm8150/commit/4927d31bfc82f7effd1a2474081a2d32b3dc0889 {"lib: zstd: define UINTPTR_MAX"}

[1]: https://github.com/facebook/zstd/commits/v1.5.4

+++ lib/zstd/compress/zstd_cwksp.h
@@ -1,5 +1,6 @@
+/* SPDX-License-Identifier: GPL-2.0+ OR BSD-3-Clause */
 /*
- * Copyright (c) Yann Collet, Facebook, Inc.
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
  * All rights reserved.
  *
  * This source code is licensed under both the BSD-style license (found in the
@@ -451,7 +452,7 @@ MEM_STATIC void ZSTD_cwksp_clean_tables(ZSTD_cwksp* ws) {
     assert(ws->tableValidEnd >= ws->objectEnd);
     assert(ws->tableValidEnd <= ws->allocStart);
     if (ws->tableValidEnd < ws->tableEnd) {
-        ZSTD_memset(ws->tableValidEnd, 0, (BYTE*)ws->tableEnd - (BYTE*)ws->tableValidEnd);
+        ZSTD_memset(ws->tableValidEnd, 0, (size_t)((BYTE*)ws->tableEnd - (BYTE*)ws->tableValidEnd));
     }
     ZSTD_cwksp_mark_tables_clean(ws);
 }
